Successful implementation of the SEO Page Rank set of stats:
		'field' + repr(1):['googlepagerank', 'integer'],
                'field' + repr(2):['googlebacklinks','integer'],
                'field' + repr(3):['googlesearchtermhits', 'integer'],
                'field' + repr(4):['yahooinlinks', 'integer'],
                'field' + repr(5):['yahoosearchtermhits', 'integer'],
                'field' + repr(6):['alexarank', 'integer'],
                'field' + repr(7):['alexainlinks', 'integer'],
                'field' + repr(8):['bingsearchtermhits', 'integer'],

These are stored in a a dict, as defined in the schema above. You will need a BOSS APP ID Key, a BING APP ID, and a willingness to possibly abuse some T&C on the alexa stuff.  You really should get an Alexa account if you want to use this for commercial use.

Also changed, spongesite.conf has been removed.  Now it exists as spongesite.conf.example.  To use it, rename it to spongesite.conf and configure for your project.  Of course, I did not include my APP IDs or developer API keys, so you will need to get yours.  Sorry for the slight inconvenience.

Minor documentation changes.